Move restartRequired unit tests to their own file
